greybus-spec: interfaces are ATTACHED/DETACHED before SVC hello
This is the first patch of several whose purpose is to integrate the
hardware model into the preconditions and postconditions of Operations
defined later on in the spec, as well as describe the order of
Operations a bit better.

Start at the beginning: require that the SVC shall make sure all
Interface States are ATTACHED or DETACHED -- i.e., determine DETECTED
and potentially ORDER for each -- before sending the SVC Hello request
to the AP.

This lets us begin building out the Interface Lifecycle from a
well-defined place.

Don't use the protocol version operation for this -- that might go away.

There's remaining "magic" where it's not specified how the Interface
States get set to their initial values in hardware_model.rst.  We can
live with that for now.

Reviewed-by: Johan Hovold <johan@hovoldconsulting.com>
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>