greybus-spec: svc: send Module Inserted after Hello
We've got a spec hole where the Module Inserted operations currently get
sent when there is an actual insertion event. That doesn't cover boot.

Add language specifying that the SVC shall send these operations to the
AP after SVC Hello. Write this in terms of the Lifecycle States that we
have previously had the SVC set to ATTACHED or DETACHED before Hello got
sent, moving the definition of an Interface ID to the hardware model to
keep things well-defined.

This is a little ugly right now. A picture would help.

Reviewed-by: Johan Hovold <johan@hovoldconsulting.com>
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>