greybus-spec: teach Module Inserted the hardware model
Specify the exact Interface State of hot-plugged modules by:

- saying that all the Interfaces States are ATTACHED
- saying how ORDER works (first is primary, any others are secondary)
- add links to the placeholder sections where pictures will be drawn

Since ORDER is the only thing we don't know when an Interface is
ATTACHED, specifying this lets the AP know the exact Interface State of
everything the Module is attached to.

Additionally mandate that if NO_PRIMARY_INTERFACE is set, then ORDER is
ORDER_SECONDARY -- otherwise, the Interface isn't ATTACHED (which
doesn't let ORDER be ORDER_UNKNOWN).

Reviewed-by: Sandeep Patil <sspatil@google.com>
Reviewed-by: Johan Hovold <johan@hovoldconsulting.com>
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>