SW-4379: hardware_model: add AP Module requirements
More prep work for teaching connection quiescing the hardware model (and
for later patches).

The AP is special. We need to be able to refer to the "interface blocks"
that are connected to the AP by name, we need to know that they are
different than the other interface blocks, and we need to say that the
AP can do things to the SVC that other modules cannot.

Add a short section to the hardware model to cover all this. It will
eventually need changes to support UFS and other non-Greybus use cases, but
we're not going to deal with those yet.

Reviewed-by: Johan Hovold <johan@hovoldconsulting.com>
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>