SW-4379: hardware_model: define "WAKE pulse" terms
In order to define the SVC Interface Activate operation, we need a
way to refer to a wake pulse (and the "cold boot" pulse time) that maps
to the hardware model. What we have now is not enough to do that.

Fix up the WAKE sub-state section to add what we need, and update the
glossary so these terms can be used from other chapters.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>