SW-4379: intf_info: move a section around
Prep work for adding the interface's DME attributes we care about to
this section.

The Data Requirements section in this document is top-level rather than
specific to just the Manifest. That's a problem because its restrictions
are manifest-specific and don't make sense when applied to DME
attributes. Move that section down under Manifest so it doesn't apply to
the various Ara-specific DME attributes we'll add later.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>