SW-4379: SVC: add Interface V_SYS Enable/Disable
Replace the single Greybus SVC Interface Power State Set operation with
a pair of new operations, Interface V_SYS Enable and Interface V_SYS
Disable.

Rewriting this operation in terms that map directly to the hardware
model makes it easier to define and contextualize with the rest of the
specification.

Whether to have two operations or a single operation which takes an
"on/off" parameter is a matter of taste. Johan wanted to do it this way,
and I don't mind.

Reviewed-by: Johan Hovold <johan@hovoldconsulting.com>
Reviewed-by: Sandeep Patil <sspatil@google.com>
Suggested-by: Johan Hovold <johan@hovoldconsulting.com>
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>