SW-4379: SVC: add Interface UNIPRO Enable/Disable
We'll need these for the new interface boot sequence. Make sure the
operation definitions are integrated with the hardware model.

There's one important difference in the error handling in the UNIPRO
operations versus those for V_SYS and REFCLK: we can't make as many
guarantees about what happened to UNIPRO if the request failed, as in
practice this has to go through the switch, and we don't want to make as
many reliability assumptions about the SVC's communication path to it.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>