SW-4379: SVC: add Interface Activate
All the pieces are in place to define the Operation that the AP uses to
"activate" an Interface, after supplying power and clock to it.

The result of this operation, if it is successful, is enough information
for the AP to know if the Interface can speak Greybus, or UniPro but not
Greybus, or no UniPro whatsoever -- i.e., it knows the INTF_TYPE, and
can now proceed to enumerate the interface when applicable.

This is a substantial change from the existing intf_hotplug() operation,
in that the process is AP driven, rather than the SVC attempting to
activate interfaces autonomously and telling the AP afterwards when it
is done.

It is an important change to do it this way, not least because the AP
controls the power budget, and needs to drive this process in order to
conserve power by not activating modules when available energy is low.

This sequence also directly specifies which party initiates UniPro link
startup: the Interface.

In all, having this operation well defined closes several glaring spec
holes from the old way of doing things, and will result in more reliable
behavior on our current platform.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>