SW-4379: SVC: add Interface Resume
This operation is needed to take an interface which was SUSPENDED back
to ENUMERATED. To define it, we need to extend the constraints for how
the UNIPRO sub-state changes values, and add a new SVC response status
code.

The "resume sequence" in this operation is pretty much what happens
during the "activation sequence" in an Interface Activate
Operation. They both have this basic structure:

1. The SVC sends a wake pulse to the Interface.
2. The Interface wakes up and initializes itself.
3. The SVC and Interface make the UniPro link operational again.
4. When the link is operational, the Interface signals readiness for
Greybus communication by writing to the mailbox.
5. If anything goes wrong, the SVC notices.
6. The SVC tells the AP what the results were.

The main changes from activate to resume are:

- The wake pulse is shorter than the cold boot threshold, so the
interface just wakes up instead of getting reset completely.

- We already know (and the SVC checks to make sure) INTF_TYPE is
IFT_GREYBUS, because unless the AP is misbehaving, the Interface
was SUSPENDED earlier, which by definition means it had to be a
Greybus Interface.

- The SVC starts off the process of making the link operational, not
the Interface.

- We take UniPro out of Hibernate instead of doing a complete link
startup sequence.

- Some other preconditions and postconditions for the Interface
State are different.

The differences are significant enough that it's better to make this
into its own operation, rather than try to shoehorn in an "activation
type" parameter into the Interface Activate Request payload, which would
just make a mess.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>