SW-4379: Control: teach overview the HW model
This is pretty old text which is quite confusing. Update it. Delete the
bit about bundles as that's not important here, and rephrase the rest to
integrate with the definitions in the rest of the specification.

Along the way, introduce two new glossary entries we can use
consistently from now on:

- Control Connection: the Greybus Connection for a Control CPort

- Control CPort: the CPort (with ID zero) that a Greybus Interface
  implements which responds to Control Protocol Operation Requests

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>