SW-4379: Control: teach Connected the HW model
It's unfortunate that we don't have a representation for bundle state in
the hardware model. It's going to make this and some subsequent patches
more vague than they should be. But we'll do the best we can for now.

Add language to the Control Connected Operation that describes when in
the Interface Lifecycle the AP is allowed to use it. Also add links to
later chapters which will define the connection management
procedure. Also make some effort to specify that the AP has to remember
which connections are established.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>