SW-4379: Control: teach Get Manifest Size the HW model
The Interface Lifecycle allows the AP and Interfaces to coordinate
changes to Interface Manifests using these requirements:

      - Interfaces shall not change their current Manifest after sending
        it to the AP, so long as they were in the ENUMERATED Lifecycle
        State when it was retrieved, and haven't entered another
        Lifecycle State since then.

      - Interfaces may change Manifests if they enter the MODE_SWITCHING
        Lifecycle State, and can request re-enumeration when exiting
        MODE_SWITCHING.

Controlling this explicitly lets the transitions into and out of
MODE_SWITCHING clean up and re-acquire any resources (like Greybus
Connections and Bundles, and UniPro Switch-internal resources) related
to the current Manifest.

Current mechanisms for firmware update do not, for example, allow the AP
to properly close all Greybus Connections to the Interface in a way that
ensures their CPorts are reset for later use. The new firmware update
design relies on MODE_SWITCHING to handle this properly.

Start building this feature into the Control Protocol:

    - Require that Interfaces pick a Manifest size when handling a Get
      Manifest Size request.

    - Require that this size shall not change while the Interface is
      being enumerated, and for as long as the Lifecycle State stays
      ENUMERATED afterwards.

We'll do the same sort of thing for the Get Manifest Operation next.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>