SW-4379: Control: teach Get Manifest the HW model
In particular, make sure that the Interface is not allowed to change the
Manifest as long as it's ENUMERATED, along with related constraints for
the Manifest's size.

Also give the way for the Interface to guarantee to the AP that as long
as it remains ENUMERATED, the information in the Manifest will be good.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>