SW-4379: Control: teach Disconnecting the HW model
The Control Disconnecting Operation currently says the Interface will
"cease transmission on the CPort" after receiving the request, except
for responses to requests it had already received.

That won't let it respond to Ping requests, which it has to for the
connection closure procedure to work.

Fix that up, teach the operation the hardware model, and reorganize it
so that requirements associated with the AP are in the request section,
and requirements associated with the Interface are in the response
section, for readability / random access discovery.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>