SW-4379: Control: teach Disconnected the HW model
Write the operation in way that makes it clear that the Greybus
Connection is what is closed now -- the UniPro CPort might technically
still be disconnected. That can also be disconnected by the SVC later,
so write that down, too.

The operation is also missing the critical requirement that the
Interface needs to make the CPort usable for transmitting and receiving
data before sending the result. Without that, connection closure will
not work. Add the requirement.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>