SW-4379: glossary: move "Interface" from Frame to Module
"Interface Block" has previously moved to live on the Frame instead of
the Module.

Now, move the location of the "Interface" from the Frame to the Module.

The AP enumerates Greybus Modules by communicating with the Interfaces
attached to each Interface Block -- the Interface is "inside" the
Module.

Additionally, the Bundles implemented by a Module, which represent
devices inside the Module, are children of the Interface, so it really
doesn't make sense to represent the Interface as part of the Frame.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>