SW-4379: hardware_model: do minimal "Interface" re-work
Do a basic rework of the hardware model to define an Interface
abstraction.

This involves making sure the Interface State refers strictly to
components within the Frame, to make a clean boundary. Doing that means
we have to define a Switch as well as adjust some other technicalities
related to how Interfaces communicate with each other and with the
Frame.

The result is not perfect, but is at least a step in the right
direction, and will be needed by later patches which define the
Interface Lifecycle over Interfaces, rather than Interface States.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>