SW-4379: special: do minimal "Interface" re-work
Make some improvements to the Special Protocols to use "Interface"
instead of "Interface State", now that we've got the basics in place to
refer to Interfaces inside of Modules, as opposed to Interface States
inside the Frame.

Like the previous patch, this is not perfect, but it is a step in the
right direction. We will need to carefully re-read this section after
finishing the hardware model "Interface" definition and make additional
changes as necessary.

Happily, in many cases, the Special Protocols already used Interface
"correctly", i.e. the language in this chapter anticipated the changes
to the hardware model. That's a good sign that we're getting the name
"Interface" right.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>