SW-4379: SVC: canonicalize and add missing ref targets
Referring to definitions in the SVC protocol isn't so pleasant right
now, and we're about to have to do a lot of that.

Most of the ref targets in the specification use dashes (-) to separate
words, not underscores (_). Some SVC operations have ref targets, and
others don't.

Fix things up. Get the SVC targets to match the dash usage, updating all
users and making sure no warnings due to missing a user are generated
when building the spec. Add missing ref targets to operations where
necessary.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>