SW-4379: Control: canonicalize and add missing ref targets
Just like the SVC protocol, the Control Protocol is using a ref target
format that's not consistent with the rest of the spec, and not all
operation definitions have ref targets.

Fix that up as well, making sure no warnings due to missing an update to
a :ref: are generated.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>