SW-4379: fix definition of "is being enumerated"
Defining when an Interface "is being enumerated" in a consistent way is
important for subsequent patches which describe enumeration.

Right now, the definition is wrong: it misses that an Interface is being
enumerated when exiting MODE_SWITCHING, too. It also should be defined
in the section which introduces the ENUMERATED state in
hardware_model.rst, so it's easier to refer to from both Protocol
definitions and the lifecycles chapter.

Fix the definition and move it to hardware_model.rst.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>