SW-4379: lifecycles: add eject procedure
This one is straightforward.

Note that we can handle the case that the Module removal occurs after
the implementation-defined duration in the error handling language for a
forcible removal: if a Module is forcibly removed, but all the
Interfaces were OFF, nothing bad happened, and it will be easy to define
a recovery procedure. There will be nothing to do from a Greybus
perspective than inform the AP in the usual way, with an SVC Module
Removed Operation.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>