SW-4379: Add float modifiers to figtable directive
Adds a new optional argument, 'loc', to the figtable directive that allows the location modifiers in the LaTeX float package to be passed in from Sphinx. If 'loc' is omitted, LaTeX has complete control over the placement of floats such as tables and figures.

See the CTAN float package documentation for descriptions of the location modifiers.

Note: This patch only applies to PDF output.

Signed-off-by: Rob Johnson <gtgprods@gmail.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>