SW-2013: greybus-spec: Add Firmware Management Protocol specifications
This patch defines the Firmware Management Protocol and lays out how the
Interface Manifest should look like for Firmware Management Bundle.

The Firmware Management Protocol can be used by the AP to load an
Interface Firmware on an Interface and prepare its Interface State to
enter the MODE_SWITCHING Lifecycle State, or update an Interface Backend
Firmware on an Interface.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>