SW-2013: greybus-spec: Remove Control Interface Version Operation
This Operation was added earlier as there was no other protocol that
supported getting version of the firmware currently running on an
Interface.

A similar operation is now implemented by the Firmware Loader Protocol,
and it makes more sense to use that operation only from now on.

This patch removes the Control Interface Version Operation.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>