SW-4741: greybus-spec: Remove unsed control connection status values
Protocols shall use their response payload for protocol specific errors.
The control bundle error that is currently defined under
"control connection status values" should be moved to the response payload
of appropriate control protocol operations if need be.

As of today, this is unused, so remove it from the spec.

Signed-off-by: Sandeep Patil <sspatil@google.com>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>