SW-4741: svc: add status in response payload for all connection operations
All connection manangement operations require the interfaces they are
dealing with to be in a specific sub-state. If not, the operations should now
return the result in the 'status' field of the response payload. This
removes the need to piggy back on Greybus operation status for receiving
errors that are specific to SVC protocol.

Signed-off-by: Sandeep Patil <sspatil@google.com>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>