SW-4741: svc: link GB_OP_SUCCESS explicitly to the Greybus operation status
There are several operations (listed below) that have a single response
code defined for success / failure in the response payload. The requester in
this case is supposed to check with the Greybus operation status
before the response code is read. These operation don't necessarily need
a SVC protocol specific status field; however, addition of those fields
in some of the other operations causes confusion around the text that
describes responses for following operations.

Fix that by explicitly referring to the Greybus operation status in
the description of response payloads for following operations -
- SVC V_SYS Enable Response
- SVC V_SYS Disable Response
- SVC REFCLK Enable Response
- SVC REFCLK Disable Response
- SVC Unipro Enable Response
- SVC Unipro Disable Response

Signed-off-by: Sandeep Patil <sspatil@google.com>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>