SW-4741: operations: move protocol status to greybus operation status
The protocol status as defined in the spec today is actually "Greybus
Operation Status". This is interpreted by most protocols as "Protocol
operation status".

SVC protocol is uses its own protocol-specific status in response
payload for some of its operations. That makes it confusing to call
the GB_OP_* values "Protocol Status".

Rename this to Greybus Operation Status and move the section under
Operations

Signed-off-by: Sandeep Patil <sspatil@google.com>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>