SW-1996: greybus-spec: Add a TimeSync Interface Lifecycle description
This patch adds a TimeSync Interface Lifecycle description.

The Interface state TIME_SYNCING represents the activity of synchronizing
frame-time to an Interface. It is valid to enter TIME_SYNCING only from the
ENUMERATED state and valid to exit back only to the ENUMERATED state from
TIME_SYNCING.

For the duration of TIME_SYNCING the Interface is owned by the TIME_SYNCING
state and it is invalid to generate a Wake signal or any other signal that
could try to take control of the relevant wake line associated with an
Interface.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Reviewed-by: Vaibhav Hiremath <vaibhav.hiremath@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>