SW-1996: greybus-spec: Replace "wake-detect" with "wake" globally
In EVT 2.0 the wake and detect functionality has been split to separate
pins, as a result continuing to refer to pins as "wake-detect" pins is
factually inaccurate. For the purposes of time-sync only the "wake"
component is of any relevance and arguably naming operations and concepts
in terms of "wake-detect" in the specification is an inappropriate leakage
of the hardware details of specific platforms and implementations into the
bus specification. This patch therefore renames "wake-detect" to "wake" for
the relevant set of patches and additionally cleans up one or two labels
which persist in having an abbreviated "wd" contained within.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Reviewed-by: Alex Elder <elder@linaro.org>
Reviewed-by: Vaibhav Hiremath <vaibhav.hiremath@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>