SW-3418: greybus-spec: Move Interface feature byte location
During review of a manifesto change Joel Porquet raised the point that
having the feature byte of an Interface located directly after the 'type'
field means that the Manifest Descriptor Header for an Interface Descriptor
is different than an Manifest Descriptor Header for non-Interface
Descriptors. This will lead to some special case processing of Interface
Descriptor headers that can be avoided by moving the 'features' byte to one
of the Interface Descriptor specific pad bytes later in the Interface
Descriptor header.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Suggested-by: Joel Porquet <porquet_joel@projectara.com>
Cc: Johan Hovold <johan@hovoldconsulting.com>
Reviewed-by: Vaibhav Hiremath <vaibhav.hiremath@linaro.org>
Acked-by: Joel Porquet <porquet_joel@projectara.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>