SW-4988: spi: add new continue bit to rdwr bitmask
When a SPI transfer needs to be split by more than one greybus spi
transfer operation, we need to indicate it so the controller can
handle the chip select lines correctly.

Add the new bit to indicate it and move the bit description to a
proper table.

Since this field now have more meaning than just read/write flag
rename it from rdwr to xfer_flags

Signed-off-by: Rui Miguel Silva <rui.silva@linaro.org>
Reviewed-by: Viresh Kumar <viresh.kumar@linaro.org>
Reviewed-by: Joel Porquet <porquet_joel@projectara.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>