SW-5372: SVC: fix Interface Mailbox Event MAILBOX handling
Other portions of the Greybus specification expect that the mailbox
attribute is cleared before the Interface writes to it as part of
certain transitions (e.g. reboot).

However, the definitions of the operations that handle them aren't
requiring the SVC to clear the mailbox after the write is detected, so
the spec is inconsistent.

Fix that for the Interface Mailbox Event Operation. Subsequent patches
will fix Interface Activate and Interface Resume.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Jeffrey Carlyle <jcarlyle@google.com>
Acked-by: David Lin <dtwlin@google.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>