SW-5372: SVC: fix Interface Activate MAILBOX handling
The SVC needs to clear the mailbox after reading it, as other places in
the spec assume the mailbox is clear before the Interface writes to it.

Tighten up the Interface Activate request language to make that
clear. Leave the error case alone. If we ever add a non-peer SVC DME
Read Operation, the AP can pull out the invalid MAILBOX value
afterwards, possibly to allow for IFT_GREYBUS2 or some other extensions
with an old SVC.

This also fixes a bug where INTF_TYPE is not explicitly set as part of
this sequence when it's IFT_GREYBUS.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Jeffrey Carlyle <jcarlyle@google.com>
Acked-by: David Lin <dtwlin@google.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>