SW-2700: SVC: improve Disconnecting language
The last sentence in the Disconnecting Response section is hard to
parse. Reorder it to make it more readable.

The words "request" and "response" weren't capitalized in the
surrounding text, so I'm leaving them lower case here. Those words get
often and we're not consistent about whether we capitalize them. I'll
leave cleaning that up to a future find/replace patch.

Reported-by: Mitchell Tasman <tasman@leaflabs.com>
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>