SW-2700: rename device class files and fix inclusion
Rename the device class files from .txt to .rst, which is the usual
restructured text file extension (and also allows editors to recognize
the file type).

Instead of using the "include" directive, which causes many warnings
during compilation, use a local toctree. The advantage for the HTML
output is that each device class will be in a separate page, thus easing
the navigation of the device class section. The PDF output should remain
unchanged.

Signed-off-by: Joël Porquet <joel@porquet.org>
Tested-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>