SW-4763: uart: define the receive credits request operation
Add the receive credits operation. This operation is initiated
by the firmware to notify the peer that data has been processed
and more space is available in its internal output buffers.

The peer should keep count of data sent to the firmware and received
credits. When the count approaches the firmware limit, the peer should
wait for more credits before sending another data message.

Signed-off-by: Axel Haslam <haslam_axel@projectara.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>