SW-4379: SVC: close Disconnecting spec hole
There's an issue with the wording of the SVC Disconnecting Operation
which is a consequence of there being no ordering specified by UniPro
for Messages sent on different CPorts.

In particular, it's possible for this to happen:

    1A. AP sends a request on some Connection, C
    2A. AP sends a Disconnecting request about C (i.e., AP starts
        connection closure for C)
    3A. UniPro Message from 2. overtakes Message from 1. on the network

Here's another thing that could happen:

    1B. AP sends a request on some Connection, C
    2B. AP sends a Disconnecting request about C (i.e., AP starts
        connection closure for C)
    3B. A CPort arbiter within the UniPro IP's Transport Layer queues
        Message from 2. before 1.

After either of those, the result could then be:

    4. Interface receives Disconnecting request and sends a response
    5. Interface receives Message from 1.

What's an Interface to do? The notion of either side sending a Message
on one CPort "after" a Message on another is not well-defined, but we
need to make sure the path for that Connection can still be cleared out
as part of Connection Management.

To handle situations like this, make it well-defined that the Interface
immediately discards the Message it receives in step 5 (unless it's a
Ping request). This will keep the pipe clear for the AP to subsequently
send its Ping requests as it continues to close the Connection.

Reported-by: Mitchell Tasman <tasman@leaflabs.com>
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Johan Hovold <johan@hovoldconsulting.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>