SW-4379: fix suspend and power_down procedures
There are bugs in the specification for the connection closure
procedures for the power_down and suspend edges in the Interface
Lifecycle.

In particular, both of them specify that the SVC Connection Destroy
Operation happens after taking the Module side of the link to Hibernate
state. That's impossible -- destroying a Connection involves DME peer
writes to the Module's Interface, which are ignored (and will fail) when
the link is hibernated.

Fix that by moving the connection closure epilogue to before the
hibernate entry. That's a little weird, but it keeps things well-defined
in terms of when the connection is torn down and the AP CPorts are
reset.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Johan Hovold <johan@hovoldconsulting.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>