SW-4379: Control: Interface transmission for "normal" Connections
The Greybus spec doesn't make it clear when it's OK to transmit data or
Flow Control Tokens as Greybus Connections are established and closed.

Fix that for non-Control Connections:

    1. It's never OK to transmit any Segments ("empty" Segments with FCT
       bit set, or Segments with L4 payload) before receiving a Control
       Connected for the CPort.

    2. After Control Connected, it's OK (and the right thing to do) to
       transmit FCTs when E2EFC is enabled; other transmission is OK too
       provided it doesn't contradict the UniPro spec or the Greybus
       Protocol.

    3. After Disconnecting, it's only OK to respond to existing
       requests, respond to later Ping requests, or send Flow Control
       Tokens (when E2EFC is enabled).

    4. After Disconnected, it's never OK to transmit any Segments.

It's safe to write it this way now that we know the Manifest doesn't
contain a CPort Descriptor for the Control CPort.

Put the legalese for when it's OK or not OK to transmit into a section
in source/connection.rst, as it's going to be reused for the AP's side
of a Connection later.

Reported-by: Mitchell Tasman <tasman@leaflabs.com>
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Johan Hovold <johan@hovoldconsulting.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>