SW-4586: Bootrom: fix up and freeze Protocol Version
We need to freeze the Bootrom Protocol as part of incorporating it into
the new hotplug design's mode switching.

To that end, stop using the |gb-major| and |gb-minor| substitution
definitions, which can change with later versions of the
specification. Replace them with the current 0.1 version of the
specification. They won't change again.

Sneak in a little capitalization fix while we're here and have to
reflow the paragraph, anyway.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Alex Elder <elder@linaro.org>
Reviewed-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>