SW-4586: Bootrom: fix up and freeze Firmware Size
We only have one implementation of this protocol, and it's the ES3 boot
ROM. It will supply the value 2 in the boot stage parameter to the
Firmware Size request, since it's looking to pull down the stage 2
loader.

Make that permanent now, deleting the table which defined other stages,
and fix up other issues with the Operation definition.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Alex Elder <elder@linaro.org>
Reviewed-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>