SW-4586: define "Legacy Mode Switch" for ES3 boot ROM
The ES3 bridge ASIC boot ROM implements the Bootrom Protocol, which
violates assumptions and requirements made elsewhere in the Greybus
Specification which support Connection Management and the MODE_SWITCHING
Interface Lifecycle State.

We still need to support it, though.

"Supporting it" in a Greybus spec context means:

    1. Adding enough restrictions to any Bootrom Protocol
       implementations to make them "work the same" as the ES3 boot ROM

    2. Stuff enough quirks into the spec to let the AP tell apart a boot
       ROM from a "modern" Greybus Interface *before* it's enumerated.

    3. Make exceptions to the Interface's required behavior when it
       implements this Protocol, to allow it to boot the next stage
       without going through MODE_SWITCHING first (i.e., allow it to
       change its Manifest while technically remaining ENUMERATED-ish)

    4. Make recommendations to the AP to try to clean up the mess left
       behind by a MAILBOX write outside of the "boot", "resume", and
       "ms_exit" Interface Lifecycle transitions

To add restrictions making any Bootrom Protocol implementations "work
like" the ES3 boot ROM, require that if the Bootrom Protocol is
implemented, it's the only protocol implemented (other than
Control). That makes sure there are no other possible open Connections
to consider for Legacy Mode Switch.

In order to let the AP detect that it's dealing with the boot ROM:

    - Reserve the Initialization Status attribute values currently used
      by the boot ROM for implementations of the Bootrom Protocol

    - Require any Interface that implements the Bootrom Protocol to set
      the Initialization Status attributes to one of those reserved
      values.

We make several exceptions to the Interface's allowed behavior, mostly
to allow it to boot the next stage without its existing Connections
being closed, and to set the MAILBOX again even though the Interface
Lifecycle says it shouldn't. In return, we require that the new
personality not implement the Bootrom Protocol.

We also make some vague and fuzzy recommendations that the AP try to
clean up after the Interface's mess in terms of tearing down resources
allocated for the open Connections.

Complete the shoe-horn by requiring that UniPro E2EFC be disabled for
those Connections, so we don't have to deal with any "empty" Segments
carrying FCTs in flight. That basically means that the final "Ready to
Boot" Operation is semantically equivalent to the last Ping Operation
after Connection Quiescing in the normal Connection closure sequences,
at least on our current implementation.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>