SW-4379: Correct Inconsistencies in Lifecycle Transition Descriptions
This patch makes the descriptions for the enumerate, resume and ms_exit
transitions consistent with each other W.R.T. creating a route and
establishing a Control Connection between the AP and an Interface.
Specifically, enumerate and resume require the AP to recreate a route
between the AP and Interface while ms_exit does not. All three
transitions require the establishment of a Control Connection.

Enumerate, Mode Switch Exit, and Resume sections as well as the Control
Connection Establishment Sequence section were updated.

Signed-off-by: Rob Johnson <gtgprods@gmail.com>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>