SW-4379: Align ENUMERATED and Get Manifest Response Descriptions
The description for the ENUMERATED Lifecycle State indicates an
Interface's Manifest must be successfully parsed before the Interface
transitions to the ENUMERATED state. However, the description for the
Greybus Control Get Manifest Response states the AP only needs to
receive the Manifest to transition to the ENUMERATED Lifecycle State.

This patch brings the two descriptions into alignment by adding text to
the description for Greybus Control Get Manifest Response to indicate
the Manifest must be parsed successfully to transition to the ENUMERATED
Lifecycle State.

Signed-off-by: Rob Johnson <gtgprods@gmail.com>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>